+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# [NEXT]
+# [2.11.3]
 - Adds methods `enableGestures` and `enableKeyboard` in `gameRef`(BonfireGameInterface)
 - Adds mixin `KeyboardEventListener`.
+- Fix exception in `TiledWorldBuilder`.
 
 # [2.11.2]
 - BugFix[`BarLifeComponent`]: animate in web.